+++ lib/govspeak.rb
@@ -228,13 +228,7 @@ module Govspeak
       attachment = attachments.detect { |a| a[:content_id] == content_id }
       next "" unless attachment
 
-      attachment = AttachmentPresenter.new(attachment)
-      span_id = attachment.id ? %{ id="attachment_#{attachment.id}"} : ""
-      # new lines inside our title cause problems with govspeak rendering as this is expected to be on one line.
-      title = (attachment.title || "").tr("\n", " ")
-      link = attachment.link(title, attachment.url)
-      attributes = attachment.attachment_attributes.empty? ? "" : " (#{attachment.attachment_attributes})"
-      %{<span#{span_id} class="attachment-inline">#{link}#{attributes}</span>}
+      render_attachment_link(AttachmentPresenter.new(attachment))
     end
 
     # DEPRECATED: use 'Image:image-id' instead
@@ -263,6 +257,15 @@ module Govspeak
       lines << image.figcaption_html if image.figcaption?
       lines << '</figure>'
       lines.join
+    end
+
+    def render_attachment_link(attachment)
+      span_id = attachment.id ? %{ id="attachment_#{attachment.id}"} : ""
+      # new lines inside our title cause problems with govspeak rendering as this is expected to be on one line.
+      title = (attachment.title || "").tr("\n", " ")
+      link = attachment.link(title, attachment.url)
+      attributes = attachment.attachment_attributes.empty? ? "" : " (#{attachment.attachment_attributes})"
+      %{<span#{span_id} class="attachment-inline">#{link}#{attributes}</span>}
     end
 
     wrap_with_div('summary', '$!')
@@ -360,6 +363,16 @@ module Govspeak
     # to be iterated in the short term.
     extension('Attachment', /#{NEW_PARAGRAPH_LOOKBEHIND}\[Attachment:\s*(.*?)\s*\]/) do |attachment_id|
       %{<govspeak-embed-attachment id="#{attachment_id}"></govspeak-embed-attachment>}
+    end
+
+    # This is an alternative syntax for embedding attachments as links. This
+    # syntax is being used by Content Publisher and should be considered
+    # experimental
+    extension('AttachmentLink', /#{NEW_PARAGRAPH_LOOKBEHIND}\[AttachmentLink:\s*(.*?)\s*\]/) do |attachment_id|
+      attachment = attachments.detect { |a| a[:id] == attachment_id }
+      next "" unless attachment
+
+      render_attachment_link(AttachmentPresenter.new(attachment))
     end
 
   private