+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Remove `[embed:attachments:content-id]` this isn't used by any apps and has never worked
 * Add dependency on govuk_publishing_components
 * Add new `AttachementLink:attachment-id` extension and mark as experimental
 * Add new `Attachement:attachment-id` extension and mark as experimental